Basic implementaiton

+++ pong/paddle.h
@@ -10,7 +10,7 @@ public:
   
   void  move(int);                     // Int represents number of pixels to move the paddle by
   void  setPosition(int, int);         // Set the paddle to a specific position (x, y)
-  float reflectionAngle(int);          // returns what angle the ball should be bounced back at, depending on the ball's y position
-  bool  checkCollision(int, int, int); // returns true if a ball of a certain position and radius is impacting this paddle
+  double reflectionAngle(double);          // returns what angle the ball should be bounced back at, depending on the ball's y position
+  bool  checkCollision(int, double, double); // returns true if a ball of a certain position and radius is impacting this paddle
   sf::Sprite getSprite();              // Fetches the sprite obj associated with this paddle
 };
